Add support for configurable and grouped products
Generate fake URLs for not individually visible product variants

There are three types of products in Magento. Product, SimpleProduct and
GroupedProduct.

SimpleProduct is most often linked to a Product (which acts as its
config). SimpleProduct can be seen as a particular variant of a Product
(of a particular size and/or color).

However, there might be a SimpleProduct, which is not linked to any
Product and can be bought separately.

Before this commit, all products were synced separately. However, we
need all SimpleProducts which do have their Product to be grouped
together and synced in one record, where SimpleProducts are `nested` as
"type:variant".

GroupedProducts are synced accordingly, but instead of "type:variant",
they are nested with "type:member".