Disable validation for commit generations
When doing a full sync & commit, we want to be sure that all record
types are committed. Take for instance a case when a last grouped
product was removed from the catalog. There are no other grouped
products, so there is nothing to be committed indeed. However, we want
to make sure that remaining grouped products are removed from Luigi's
Box as well. The `validate_generation: false` parameter ensures that.